requirement.txt
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.2.10
astroid==2.4.2
azure-core==1.8.0
azure-storage-blob==12.4.0
cached-property==1.5.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
colorama==0.4.3
cryptography==3.1
defusedxml==0.6.0
distlib==0.3.0
Django==3.0.5
django-filter==2.4.0
djangorestframework==3.11.1
djongo==1.3.3
dnspython==2.0.0
exchangelib==3.2.1
Faker==4.1.1
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
isodate==0.6.0
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.5.2
mccabe==0.6.1
msrest==0.6.18
ntlm-auth==1.5.0
oauthlib==3.1.0
Pillow==8.0.1
pipenv==2020.6.2
psycopg2==2.8.6
pycparser==2.20
Pygments==2.6.1
pylint==2.5.3
pymongo==3.11.0
python-dateutil==2.8.1
python-imap==1.0.0
pytz==2020.1
requests==2.24.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
schedule==0.6.0
selenium==3.141.0
six==1.15.0
sqlparse==0.2.4
text-unidecode==1.3
toml==0.10.1
tzlocal==2.1
urllib3==1.25.10
virtualenv==20.0.24
virtualenv-clone==0.5.4
whitenoise==5.2.0
wrapt==1.12.1
